fix: increase # of emails fetched to find letters
Might help fix an issue that @martinsrna was facing where his Substack
newsletters weren't showing up in the "Letters" page. I'm guessing that
this was because there were no such Substack messages in his last 500
emails and thus the app couldn't detect them.

So, I've upped the # of messages fetched to 2500 (in batches of 100).