perf(lib/hooks/messages): don't revalidate all pages
Revalidating all `/api/messages` pages means that, for every new page
that needs to be loaded, all the previous pages have to first be
revalidated (at least, that's what I've noticed in my testing).

This makes it so that as the user continues to scroll, load times
increase exponentially. This is unacceptable so I'm disabling this
option by default. Also, it's not necessary to revalidate anyways
because I'm already locally mutating the feed data as needed.